changed variable to long unsigned int

+++ src/exif.c
@@ -557,7 +557,7 @@ static void ProcessExifDir(unsigned char * DirStart, unsigned char * OffsetBase,
         if (ShowTags){
             /* Show tag name */
             for (a=0;;a++){
-                if (a >= TAG_TABLE_SIZE){
+                if ((long unsigned int)a >= TAG_TABLE_SIZE){
                     printf("%s    Unknown Tag %04x Value = ", IndentString, Tag);
                     break;
                 }
